proposed method for kwarg sharing. Not sure if I like it though
