+++ e3nn/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.0.0-alpha0"
+__version__ = "0.0.0"
